Armbian

+++ docker/fxsupport/linux/resize.sh
@@ -8,10 +8,15 @@ if test -f /etc/apt/apt.conf.d/proxy.conf; then sudo rm /etc/apt/apt.conf.d/prox
 resize_rootfs () {
   touch /usr/bin/fula/.resize_flg
   #sh /usr/lib/raspi-config/init_resize.sh
-  sudo raspi-config --expand-rootfs
-  echo "Rootfs expanded..."
-  sudo reboot
-  exit 0
+  if [ -d "/sys/module/rockchipdrm" ]; then
+    echo "Running on RockChip. No action needed"
+    exit 0
+  else
+    sudo raspi-config --expand-rootfs
+    echo "Rootfs expanded..."
+    sudo reboot
+    exit 0
+  fi
 }
 
 if [ -f "$resize_flag" ]; then
